requirements.txt
requests==1.2.0
pyaudio
scikits.samplerate